Reimplement recovery constructors
The API changes in this commit are breaking changes.

Remove `MT::recover_from_iter` and implement `MT::recover` as taking an
`IntoIterator`. Return a `Result` instead of an `Option`.

Add a new `RecoverRngError` type that implements `std::error::Error`
when a new `std` feature is enabled.

Add `TryFrom<&[u32]>` and `TryFrom<&[u64]>` to replace the old
slice-based `recover` APIs.

Add `From` impls that take the correct fix-sized array and are
infallible.