Add tests for Mt::recover and Mt64::recover
API pairs `Mt::recover`, `impl From<[u32; N]> for Mt`, and
`impl TryFrom<&[u32]> for Mt` recover recover an RNG, but `recover`
operates on iterators whereas the `From` and `TryFrom` versions operate
on slices. These APIs are all implemented independently.

This PR duplicates the tests that already exist for the `From` and
`TryFrom` impls to also test `recover`.

This PR makes the same changes for `Mt64`.