Fix Apache 2.0 license
When forking the repository in #1, I misinterpreted a section of the
Apache-2.0 license. This section describes how to attribute copyright
using the Apache 2.0 license, but I filled in the actual copyright.

This commit undoes these changes and copies the license from:

https://www.apache.org/licenses/LICENSE-2.0.txt

This has the effect of changing all of the whitespace in the
`LICENSE-APACHE` file.

To see the substantive portions of this commit, suppress whitespace
changes.