Do not eagerly call `next_uNN` in `fill_bytes` with empty remainder
Fix a bug where random numbers are consumed from the Mersenne Twister
even when they will never be used because the remainder from the chunks
iterator in `fill_bytes` is non-empty.

This regression was introduced in:

- https://github.com/artichoke/rand_mt/pull/116

Add a basic smoke check for reproducibility with MRI's RNG pulled from
`spinoso-random` in artichoke/artichoke.