Remove benches
These benchmarks were useful while this crate was actively developed.
Now that the crate is passively maintained, remove this code to reduce
maintenance burden.

This change also removes a dependency on nightly for this project, which
allows using stable clippy and rustfmt in Rake tasks and CI, which should
lead to less frequent (and more predictable timing of) clippy breakage in
CI.